I moved the config folder to the root of the app
